Hackathon 2021: move parameter update function for adjoint iteration to helper_functions. Move iteration from core.cc into the specific adjoint solver scheme.
