+++ source/simulator/core.cc
@@ -468,7 +468,6 @@ namespace aspect
     // We need to do the RHS compatibility modification, if the model is
     // compressible or compatible (in the case of melt transport), and
     // there is no open boundary to balance the pressure.
-    // The same correction has to be applied for the adjoint equation
     do_pressure_rhs_compatibility_modification = ((material_model->is_compressible() && !parameters.include_melt_transport)
                                                   ||
                                                   (parameters.include_melt_transport && !material_model->is_compressible())
@@ -1736,9 +1735,6 @@ namespace aspect
   Simulator<dim>::
   solve_timestep ()
   {
-    // set adjoint problem to false as default
-    adjoint_problem = false;
-
     // start any scheme with an extrapolated value from the previous
     // two time steps if those are available
     initialize_current_linearization_point();
@@ -1939,7 +1935,7 @@ namespace aspect
           }
       }
 
-    // Start the principal loop over time steps. At this point, everything
+// Start the principal loop over time steps. At this point, everything
     // is completely initialized, so set that status as well
     simulator_is_past_initialization = true;
     do
@@ -1948,50 +1944,34 @@ namespace aspect
         // solving in pre-refinement.
         if (! (parameters.skip_solvers_on_initial_refinement
                && pre_refinement_step < parameters.initial_adaptive_refinement))
-
-      if (! (parameters.skip_solvers_on_initial_refinement
-             && pre_refinement_step < parameters.initial_adaptive_refinement))
-        {
-          start_timestep ();
-        }
-
-        // since the default for num_it_adjoint i 0, this won't do anything if this isn't run in adjoint mode
-        for (unsigned int i=0; i<parameters.num_it_adjoint; ++i)
           {
+            start_timestep ();
 
-            if (parameters.nonlinear_solver == NonlinearSolver::Stokes_adjoint)
-              pcout << " ^^ Adjoint iteration number " << i  << std::endl;
-
-        // Only solve if we are not in pre-refinement, or we do not want to skip
-        // solving in pre-refinement.
-        if (! (parameters.skip_solvers_on_initial_refinement
-               && pre_refinement_step < parameters.initial_adaptive_refinement))
-          {
             // then do the core work: assemble systems and solve
             solve_timestep ();
           }
 
-            // See if we have to start over with a new adaptive refinement cycle
-            // at the beginning of the simulation. If so, set the
-            // simulator_is_past_initialization variable back to false because we will
-            // have to re-initialize some variables such as the size of vectors,
-            // the initial state, etc.
-            if (timestep_number == 0)
+        // See if we have to start over with a new adaptive refinement cycle
+        // at the beginning of the simulation. If so, set the
+        // simulator_is_past_initialization variable back to false because we will
+        // have to re-initialize some variables such as the size of vectors,
+        // the initial state, etc.
+        if (timestep_number == 0)
+          {
+            const bool initial_refinement_done = maybe_do_initial_refinement(max_refinement_level);
+            if (initial_refinement_done)
               {
-                const bool initial_refinement_done = maybe_do_initial_refinement(max_refinement_level);
-                if (initial_refinement_done)
-                  {
-                    simulator_is_past_initialization = false;
-                    goto start_time_iteration;
-                  }
+                simulator_is_past_initialization = false;
+                goto start_time_iteration;
               }
+          }
 
         // Prepare the next time step:
         time_stepping_manager.update();
 
         const double new_time_step_size = time_stepping_manager.get_next_time_step_size();
 
-        // if we postprocess nonlinear iterations or adjoint stokes, this function is called within
+        // if we postprocess nonlinear iterations, this function is called within
         // solve_timestep () in the individual solver schemes
         if (!time_stepping_manager.should_repeat_time_step()
             && !parameters.run_postprocessors_on_nonlinear_iterations)
@@ -2025,8 +2005,8 @@ namespace aspect
             continue; // repeat time step loop
           }
 
-            // see if we want to write a timing summary
-            maybe_write_timing_output();
+        // see if we want to write a timing summary
+        maybe_write_timing_output();
 
         // update values for timestep, increment time step by one.
         advance_time(new_time_step_size);
@@ -2039,20 +2019,12 @@ namespace aspect
 
         const bool checkpoint_written = maybe_write_checkpoint(last_checkpoint_time,
                                                                write_checkpoint_due_to_termination);
-        }
-
-        // check whether to terminate the simulation. the
-        // first part of the pair indicates whether to terminate
-        // the execution; the second indicates whether to do one
-        // more checkpoint
-        const std::pair<bool,bool> termination = termination_manager.execute();
 
         if (checkpoint_written)
           last_checkpoint_time = std::time(nullptr);
 
         if (should_terminate)
           break;
-
       }
     while (true);
 
@@ -2065,7 +2037,6 @@ namespace aspect
     stokes_matrix_free.reset();
   }
 }
-
 
 
 // explicit instantiation of the functions we implement in this file
